steam: Fix float stat truncation, outdated macOS SDK path, and VS detection
steamshim_child.c: Fix float stat values being silently truncated to
integers. In processEvent(), SHIMEVENT_SETSTATF/GETSTATF was assigning
float values as:

    event.fvalue = (int)*((float *)buf);

The erroneous (int) cast discarded the fractional part before assigning
to the float field. This meant stats like stat_online_minutes_played
(a float) were always rounded down to whole numbers. Remove the cast so
the float value is assigned directly.

steamshim_parent/Makefile: Update macOS redistributable library path
from osx32 to osx. Valve renamed the directory when they dropped 32-bit
macOS support in Steamworks SDK ~1.42 (2018). Building the Steam shim
for HOST=osx would fail to link because the path no longer exists in
any recent SDK.

steamshim_parent/build.bat: Add support for Visual Studio 2017, 2019,
and 2022. Previously the script only detected VS2015 (VS140COMNTOOLS).
Now it first tries vswhere.exe (available since VS2017, the standard
Microsoft-recommended discovery mechanism) to locate the latest
installed VS with C++ tools. Falls back to the legacy VS2015 check if
vswhere is not found.

+++ source/glest_game/steamshim/steamshim_child.c
@@ -268,7 +268,7 @@ static const STEAMSHIM_Event *processEvent(const uint8 *buf, size_t buflen) {
     case SHIMEVENT_SETSTATF:
     case SHIMEVENT_GETSTATF:
         event.okay = *(buf++) ? 1 : 0;
-        event.fvalue = (int)*((float *)buf);
+        event.fvalue = *((float *)buf);
         buf += sizeof(float);
         strcpy(event.name, (const char *)buf);
         break;
